requirements.txt
prometheus-client
requests
psutil
urllib3
uplink
pytest
pytest-cov
python-dotenv